+++ Overlay/TerrainData.cs
@@ -25,15 +25,15 @@ namespace Overlay
 
         }
 
-        private readonly CellMap<float> heightRatios;
+        private readonly CellMap<double> heightRatios;
 
         private TerrainData(CelestialBody body, int gridLevel)
         {
             if (body.pqsController == null) { throw new ArgumentException("Body doesn't have a PQS controller"); }
-            heightRatios = new CellMap<float>(gridLevel, c => (float)(body.pqsController.GetSurfaceHeight(c.Position) / body.pqsController.radius));
+            heightRatios = new CellMap<double>(gridLevel, c => (body.pqsController.GetSurfaceHeight(c.Position) / body.pqsController.radius));
         }
 
-        public float GetHeightRatio(Cell cell)
+        public double GetHeightRatio(Cell cell)
         {
             return heightRatios[cell];
         }
